make: delete XDIFF_LIB, add xdiff to LIB_OBJS
In a future patch series the 'xdiff' Rust crate will be added. Delete
the creation of the static library file for xdiff to avoid a name
conflict. This also moves toward the goal of Rust only needing to link
against libgit.a.

Changes to Meson are not required as the xdiff library is already
included in Meson's libgit.a.

Signed-off-by: Ezekiel Newren <ezekielnewren@gmail.com>