xdiff: change type of xdfile_t.changed from char to bool
The only values possible for 'changed' is 1 and 0, which exactly maps
to a bool type. It might not look like this is the case because
matches1 and matches2 (which use to be dis1, and dis2) were also char
and were assigned numerical values within a few lines of 'changed'
(what used to be rchg).

Using NONE, SOME, TOO_MANY for matches1[i]/matches2[j], and true/false
for changed[k] makes it clear to future readers that these are
logically separate concepts.

Best-viewed-with: --color-words
Signed-off-by: Ezekiel Newren <ezekielnewren@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>