read-cache: add/read-cache respect submodule ignore=all
Submodules configured with ignore=all are now skipped during add operations
unless overridden by --include-ignored-submodules and the submodule path is
explicitly specified.

A message is printed (like ignored files) guiding the user to use the
--include-ignored-submodules flag if the user has explicitely want to update
the submodule reference.

The reason for the change is support submodule branch tracking or
similar and git status state nothing and git add should not add either.
The workflow is more logic and similar to regular ignored files even
the submodule is already tracked.

The change opens up a lot of possibilities for submodules to be used
more freely and simular to the repo tool. A submodule can be added for many
more reason and loosely coupled dependencies to the super repo which often
gives the friction of handle the explicit commits and updates without
the need for tracking the submodule sha1 by sha1.

Signed-off-by: Claus Schneider(Eficode) <claus.schneider@eficode.com>