remembering-renames/sparse-checkout: renumber lists starting at zero
In 52751000bc9 (doc: add some missing technical documents, 2025-10-02),
some documents were added to the default set of files to be rendered.

Unfortunately, this led to CI failures since Oct 10 (see
https://github.com/git/git/actions/runs/18418984731), the eight most
recent pushes of Git's `master` as of time of writing.

Due to the way Git's automation is set up, it is quite difficult to get
to the actual problems, as some known-bogus warnings get suppressed in
the check that fails the job, but those warnings are _not_ suppressed in
the failed jobs' logs. The actual problems look like this:

  asciidoc: WARNING: remembering-renames.adoc: line 13: list item index: expected 1 got 0
  asciidoc: WARNING: remembering-renames.adoc: line 15: list item index: expected 2 got 1
  asciidoc: WARNING: remembering-renames.adoc: line 17: list item index: expected 3 got 2
  asciidoc: WARNING: remembering-renames.adoc: line 20: list item index: expected 4 got 3
  asciidoc: WARNING: remembering-renames.adoc: line 23: list item index: expected 5 got 4
  asciidoc: WARNING: remembering-renames.adoc: line 25: list item index: expected 6 got 5
  asciidoc: WARNING: remembering-renames.adoc: line 29: list item index: expected 7 got 6
  asciidoc: WARNING: remembering-renames.adoc: line 31: list item index: expected 8 got 7
  asciidoc: WARNING: remembering-renames.adoc: line 33: list item index: expected 9 got 8
  asciidoc: WARNING: remembering-renames.adoc: line 38: section title out of sequence: expected level 1, got level 2
  asciidoc: WARNING: sparse-checkout.adoc: line 17: section title out of sequence: expected level 1, got level 2
  asciidoc: WARNING: sparse-checkout.adoc: line 928: list item index: expected 1 got 0
  asciidoc: WARNING: sparse-checkout.adoc: line 931: list item index: expected 2 got 1
  asciidoc: WARNING: sparse-checkout.adoc: line 951: list item index: expected 3 got 2
  asciidoc: WARNING: sparse-checkout.adoc: line 974: list item index: expected 4 got 3
  asciidoc: WARNING: sparse-checkout.adoc: line 980: list item index: expected 5 got 4
  asciidoc: WARNING: sparse-checkout.adoc: line 1033: list item index: expected 6 got 5
  asciidoc: WARNING: sparse-checkout.adoc: line 1049: list item index: expected 7 got 6

The problems fall into two categories. The first one is that there are
two lists that start at zero instead of 1, which is unexpected for
AsciiDoc. In the same spirit as fef11965da8 (Renumber list in
api-command.txt, 2012-12-15), let's simply renumber those lists. The
second category will be addressed in a separate commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>